Refactoring the architecture
