+++ src/main/java/cs544/group1/project/domain/Role.java
@@ -1,0 +1,7 @@
+package cs544.group1.project.domain;
+
+public enum Role {
+	
+	ADMIN, STUDENT, CHECKER;
+
+}
