+++ src/main/java/cs544/group1/project/repo/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import cs544.group1.project.domain.User;
 
 @Repository
-public interface UserRepo extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 	
 	
 }
