--- src/main/java/cs544/group1/project/types/ReservationStatus.java
@@ -1,7 +1,0 @@
-package cs544.group1.project.types;
-
-public enum ReservationStatus {
-	
-	PENDING, ACCEPTED, DECLINED;
-
-}
