--- src/main/java/cs544/group1/project/types/UserRoles.java
@@ -1,7 +1,0 @@
-package cs544.group1.project.types;
-
-public enum UserRoles {
-	
-	ADMIN, STUDENT, CHECKER;
-
-}
